refactor(core): move error contracts to Core
